refactor(frontend): pages/consumer/hooks/ を model/ へ移設する（#1043）
A1' の3本目（最後）。規約射程内・道具射程外だった1件。

- use-entity-id-by-slug.ts を hooks/ → model/ へ（R100・中身バイト不変）
- 参照1件（PublicRecordDetailPage.tsx）の import を追随
- 空になった hooks/ を削除 → **リポジトリから hooks/ ディレクトリが消えた（0個）**

codemod は features/<slice>/hooks/ しか見ないため本件は手作業。1ファイルなので fleet 推奨どおり。

規約正本 = 01-architecture.md:99（スライス内＝pages/features/entities のサブディレクトリは
ui/model/api/lib/config の5語のみ・hook は model/）。

検証: npm run check --prefix frontend = exit 0（vitest 119 files 666 tests 含む）。

Closes #1043
