+++ frontend/src/pages/consumer/model/use-entity-id-by-slug.ts
